For BZ 62023, followup to r1822004 - Use explicit unboxing.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/tc8.5.x/trunk@1822017 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/tomcat/util/net/AprEndpoint.java
@@ -629,12 +629,12 @@ public class AprEndpoint extends AbstractEndpoint<Long> implements SNICallBack {
     @Override
     protected void releaseSSLContext(SSLHostConfig sslHostConfig) {
         Long ctx = sslHostConfig.getOpenSslContext();
-        if (ctx != null && ctx != 0) {
+        if (ctx != null && ctx.longValue() != 0L) {
             SSLContext.free(ctx.longValue());
             sslHostConfig.setOpenSslContext(null);
         }
         Long cctx = sslHostConfig.getOpenSslConfContext();
-        if (cctx != null && cctx != 0) {
+        if (cctx != null && cctx.longValue() != 0L) {
             SSLConf.free(cctx.longValue());
             sslHostConfig.setOpenSslConfContext(null);
         }
